# Recommender-Service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an image processing server.",
        "title": "Swagger Recommender-System",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "basePath": "/",
    "paths": {
        "/api/leisure/list": {
            "post": {
                "description": "handler for recommendation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "default": "\u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Request structure",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/algoStorage.RecommendRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/algoStorage.RecommendResponse"
                        }
                    },
                    "403": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    },
                    "507": {
                        "description": ""
                    }
                }
            }
        },
        "/api/tickets/": {
            "post": {
                "description": "handler for tickets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "default": "\u003cAdd access token here\u003e",
                        "description": "Insert your access token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Request structure",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.TicketsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.TicketsResponse"
                        }
                    },
                    "403": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "507": {
                        "description": ""
                    }
                }
            }
        },
        "/api/user/auth": {
            "post": {
                "description": "authorization handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "All"
                ],
                "parameters": [
                    {
                        "description": "Request structure",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.AuthResponse"
                        }
                    },
                    "401": {
                        "description": ""
                    },
                    "404": {
                        "description": ""
                    },
                    "507": {
                        "description": ""
                    }
                }
            }
        },
        "/api/user/registration": {
            "post": {
                "description": "registration handler",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "All"
                ],
                "parameters": [
                    {
                        "description": "Request structure",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.UserInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/storage.AuthResponse"
                        }
                    },
                    "404": {
                        "description": ""
                    },
                    "409": {
                        "description": ""
                    },
                    "507": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "algoStorage.RecommendRequest": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string"
                },
                "in_time": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "algoStorage.RecommendResponse": {
            "type": "object",
            "properties": {
                "arts_list": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                }
            }
        },
        "storage.AuthResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "storage.Ticket": {
            "type": "object",
            "properties": {
                "arrivalTime": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "departureTime": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "transfer": {
                    "type": "string"
                },
                "transferDepartureTime": {
                    "type": "string"
                },
                "transferTime": {
                    "type": "string"
                }
            }
        },
        "storage.TicketsRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "storage.TicketsResponse": {
            "type": "object",
            "properties": {
                "tickets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/storage.Ticket"
                    }
                }
            }
        },
        "storage.UserInfo": {
            "type": "object",
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        }
    }
}